+++ routes/api/v1/products.js
@@ -1,16 +1,24 @@
 const router = require('express').Router();
 const Products = require('../../../models/Products');
+const Registers = require('../../../models/Registers');
 const Mods = require('../../../models/Mods');
 const utils = require('../../../utils');
 const bluebird = require('bluebird');
 
 router.get('/products', (req, res) => {
-  Products.find({ regId: req.user.regId }).select('-_id -__v -regId').then((products) => {
+  Registers.findOne({ _id: req.user.regId }).select('syncRegId').then((register) => {
+    return Products.find({ regId: register.syncRegId || req.user.regId }).select('-_id -__v -regId');
+  }).then((products) => {
     res.json(products);
   }).catch(utils.handleError(res));
 });
 
-router.post('/products', (req, res) => {
+router.post('/products', async (req, res) => {
+  const register = await Registers.findOne({ _id: req.user.regId }).select('syncRegId');
+  if (register.syncRegId) {
+    return res.status(403).json({ success: false, msg: 'srv_cannot_modify_synced_reg' });
+  }
+
   const { ean, highlight, mods, ...product } = req.body;
   const update = { $set: product };
   if (!highlight) {
@@ -39,7 +47,12 @@ router.post('/products', (req, res) => {
   }).catch(utils.handleError(res));
 });
 
-router.delete('/products/:ean', (req, res) => {
+router.delete('/products/:ean', async (req, res) => {
+  const register = await Registers.findOne({ _id: req.user.regId }).select('syncRegId');
+  if (register.syncRegId) {
+    return res.status(403).json({ success: false, msg: 'srv_cannot_modify_synced_reg' });
+  }
+
   const { ean } = req.params;
   Products.deleteOne({ regId: req.user.regId, ean }).then(() => {
     return Mods.find({ regId: req.user.regId, [`eans.${ean}`]: true }).select('eans number');
